+++ fabric/src/main/java/com/lx862/jcm/mixin/minecraft/NativeImageAccessor.java
@@ -7,5 +7,5 @@ import org.spongepowered.asm.mixin.gen.Accessor;
 @Mixin(NativeImage.class)
 public interface NativeImageAccessor {
     @Accessor
-    public long getPointer();
+    long getPointer();
 }